pyproject.toml
[project]
name = "abstractivekg-nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "transformers",
    "torch",
    "spacy",
    "neo4j",
    "networkx",
    "datasets",
    "evaluate",
    "rouge_score",
    "accelerate",
    "python-dotenv"
]
